requirements.txt
django~=2.2
djangorestframework~=3.12
djangorestframework-simplejwt~=4.7
flake8~=5.0
flake8-broken-line~=0.6
flake8-isort~=5.0
flake8-return~=1.2
pep8-naming~=0.13
PyJWT~=2.1
pytest~=6.2
pytest-django~=4.4
pytest-pythonpath~=0.7
requests~=2.26
